+++ src/canvas/Canvas.ts
@@ -198,7 +198,7 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
   }
 
   /**
-   * Removes all event listeners
+   * Removes all event listeners, used when disposing the instance
    */
   removeListeners() {
     this.addOrRemove(removeListener, 'remove');
@@ -228,6 +228,7 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
       this._onMouseMove as EventListener,
       addEventOptions,
     );
+    clearTimeout(this._willAddMouseDown);
   }
 
   /**
@@ -597,11 +598,24 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
    * @param {Event} e Event object fired on mousedown
    */
   _onTouchStart(e: TouchEvent) {
-    e.preventDefault();
+    // we will prevent scrolling if allowTouchScrolling is not enabled and
+    let shouldPreventScrolling = !this.allowTouchScrolling;
+    const currentActiveObject = this._activeObject;
     if (this.mainTouchId === undefined) {
       this.mainTouchId = this.getPointerId(e);
     }
     this.__onMouseDown(e);
+    // after executing fabric logic for mouse down let's see
+    // if we didn't change target or if we are drawing
+    // we want to prevent scrolling anyway
+    if (
+      this.isDrawingMode ||
+      (currentActiveObject && this._target === currentActiveObject)
+    ) {
+      shouldPreventScrolling = true;
+    }
+    // prevent default, will block scrolling from start
+    shouldPreventScrolling && e.preventDefault();
     this._resetTransformEventData();
     const canvasElement = this.upperCanvasEl,
       eventTypePrefix = this._getEventPrefix();
@@ -612,12 +626,14 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
       this._onTouchEnd as EventListener,
       addEventOptions,
     );
-    addListener(
-      doc,
-      'touchmove',
-      this._onMouseMove as EventListener,
-      addEventOptions,
-    );
+    // if we scroll don't register the touch move event
+    shouldPreventScrolling &&
+      addListener(
+        doc,
+        'touchmove',
+        this._onMouseMove as EventListener,
+        addEventOptions,
+      );
     // Unbind mousedown to prevent double triggers from touch devices
     removeListener(
       canvasElement,
@@ -1430,7 +1446,7 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
         }
         this._fireSelectionEvents(prevActiveObjects, e);
       } else {
-        (activeObject as IText).exitEditing &&
+        (activeObject as IText).isEditing &&
           (activeObject as IText).exitEditing();
         // add the active object and the target to the active selection and set it as the active object
         const klass =