+++ src/shapes/IText/ITextBehavior.ts
@@ -387,6 +387,22 @@ export abstract class ITextBehavior<
     if (this.isEditing || !this.editable) {
       return;
     }
+    this.enterEditingImpl();
+    this.fire('editing:entered', e ? { e } : undefined);
+    this._fireSelectionChanged();
+    if (this.canvas) {
+      this.canvas.fire('text:editing:entered', {
+        target: this as unknown as IText,
+        e,
+      });
+      this.canvas.requestRenderAll();
+    }
+  }
+
+  /**
+   * runs the actual logic that enter from editing state, see {@link enterEditing}
+   */
+  enterEditingImpl() {
     if (this.canvas) {
       this.canvas.calcOffset();
       this.canvas.textEditingManager.exitTextEditing();
@@ -403,15 +419,6 @@ export abstract class ITextBehavior<
     this._textBeforeEdit = this.text;
 
     this._tick();
-    this.fire('editing:entered', e ? { e } : undefined);
-    this._fireSelectionChanged();
-    if (this.canvas) {
-      this.canvas.fire('text:editing:entered', {
-        target: this as unknown as IText,
-        e,
-      });
-      this.canvas.requestRenderAll();
-    }
   }
 
   /**
@@ -669,6 +676,9 @@ export abstract class ITextBehavior<
 
   /**
    * runs the actual logic that exits from editing state, see {@link exitEditing}
+   * Please use exitEditingImpl, this function was kept to avoid breaking changes.
+   * Will be removed in fabric 7.0
+   * @deprecated use "exitEditingImpl"
    */
   protected _exitEditing() {
     const hiddenTextarea = this.hiddenTextarea;
@@ -686,10 +696,10 @@ export abstract class ITextBehavior<
   }
 
   /**
-   * Exits from editing state and fires relevant events
+   * runs the actual logic that exits from editing state, see {@link exitEditing}
+   * But it does not fire events
    */
-  exitEditing() {
-    const isTextChanged = this._textBeforeEdit !== this.text;
+  exitEditingImpl() {
     this._exitEditing();
     this.selectionEnd = this.selectionStart;
     this._restoreEditingProps();
@@ -697,6 +707,15 @@ export abstract class ITextBehavior<
       this.initDimensions();
       this.setCoords();
     }
+  }
+
+  /**
+   * Exits from editing state and fires relevant events
+   */
+  exitEditing() {
+    const isTextChanged = this._textBeforeEdit !== this.text;
+    this.exitEditingImpl();
+
     this.fire('editing:exited');
     isTextChanged && this.fire(MODIFIED);
     if (this.canvas) {